refactor: extract WSL-aware openHostPath helper
